# Use different buffer size for every integer type