@dtolnay 2018-03-22
I released this improvement in 0.4.1.